fix frames_in_flight drop bug
The problem was when attempting to drop the frames_in_flight structure
while a frame was mid-recording. The call to
wait_for_all_frames_to_complete would block waiting for the frame's
fence, but if there was an error while _recording_ the frame, then the
frame's commands would never be submitted and the fence would never get
signalled.

The fix here is to mark an in-progress frame and exclude it from the
wait.